Multiple code refactorings
