+++ common/app/layout/FaciaCardTimestamp.scala
@@ -1,0 +1,40 @@
+package layout
+
+import cards.{MediaList, Standard}
+import com.gu.commercial.branding.Branding
+import com.gu.contentapi.client.model.{v1 => contentapi}
+import com.gu.contentapi.client.utils.{AdvertisementFeature, DesignType}
+import common.Edition.defaultEdition
+import common.{Edition, LinkTo}
+import implicits.FaciaContentFrontendHelpers.FaciaContentFrontendHelper
+import model._
+import model.pressed._
+import org.joda.time.DateTime
+import play.api.mvc.RequestHeader
+import play.twirl.api.Html
+import services.FaciaContentConvert
+import views.support._
+
+import scala.Function.const
+
+sealed trait FaciaCardTimestamp {
+  def javaScriptUpdate: Boolean
+
+  def formatString: String
+}
+
+// By default a date string, but uses JavaScript to update to a human readable string like '22h' meaning 22 hours ago
+case object DateOrTimeAgo extends FaciaCardTimestamp {
+  override val javaScriptUpdate: Boolean = true
+  override val formatString: String = "d MMM y"
+}
+
+case object DateTimestamp extends FaciaCardTimestamp {
+  override val javaScriptUpdate: Boolean = false
+  override val formatString: String = "d MMM y"
+}
+
+case object TimeTimestamp extends FaciaCardTimestamp {
+  override val javaScriptUpdate: Boolean = false
+  override val formatString: String = "h:mm aa"
+}
