+++ common/app/layout/FaciaContainerHeader.scala
@@ -6,27 +6,19 @@ import org.joda.time.LocalDate
 import org.joda.time.format.DateTimeFormat
 import services.ConfigAgent
 
-sealed trait FaciaHeaderImageType
+sealed trait FaciaContainerHeader
 
-/** TODO fancy cut out version */
-case object ContributorCutOutImage extends FaciaHeaderImageType
+case class MetaDataHeader(
+  displayName: String,
+  image: Option[FaciaHeaderImage],
+  description: Option[String],
+  dateHeadline: DateHeadline,
+  href: Option[String]
+) extends FaciaContainerHeader
 
-case object ContributorCircleImage extends FaciaHeaderImageType
-case object FootballBadge extends FaciaHeaderImageType
+case class LoneDateHeadline(get: DateHeadline) extends FaciaContainerHeader
 
-case class FaciaHeaderImage(
-  url: String,
-  imageType: FaciaHeaderImageType
-) {
-  def cssClasses: Seq[String] = Seq(
-    "index-page-header__image-wrapper",
-    imageType match {
-      case ContributorCircleImage => "index-page-header__image-wrapper--contributor-circle"
-      case ContributorCutOutImage => "index-page-header__image-wrapper--contributor-cut-out"
-      case FootballBadge => "index-page-header__image-wrapper--football-badge"
-    }
-  )
-}
+case class DescriptionMetaHeader(description: String) extends FaciaContainerHeader
 
 object FaciaContainerHeader {
   def fromSection(sectionPage: Section, dateHeadline: DateHeadline)(implicit context: ApplicationContext): FaciaContainerHeader = MetaDataHeader(
@@ -82,56 +74,4 @@ object FaciaContainerHeader {
     } else {
       None
     }
-}
-
-sealed trait FaciaContainerHeader
-
-case class MetaDataHeader(
-  displayName: String,
-  image: Option[FaciaHeaderImage],
-  description: Option[String],
-  dateHeadline: DateHeadline,
-  href: Option[String]
-) extends FaciaContainerHeader
-
-case class LoneDateHeadline(get: DateHeadline) extends FaciaContainerHeader
-
-case class DescriptionMetaHeader(description: String) extends FaciaContainerHeader
-
-object DateHeadline {
-  def cardTimestampDisplay(dateHeadline: DateHeadline): FaciaCardTimestamp = dateHeadline match {
-    case _: DayHeadline => TimeTimestamp
-    case _: MonthHeadline => DateTimestamp
-  }
-}
-
-sealed trait DateHeadline {
-  val dateFormatString: String
-
-  val dateTimeFormatString: String
-
-  // TODO add a month endpoint and then make this non-optional
-  val urlFragmentFormatString: Option[String]
-
-  val day: LocalDate
-
-  def displayString: String = day.toDateTimeAtStartOfDay.toString(DateTimeFormat.forPattern(dateFormatString))
-
-  def dateTimeString: String = day.toDateTimeAtStartOfDay.toString(DateTimeFormat.forPattern(dateTimeFormatString))
-
-  def urlFragment: Option[String] = urlFragmentFormatString map { format =>
-    day.toDateTimeAtStartOfDay.toString(DateTimeFormat.forPattern(format)).toLowerCase
-  }
-}
-
-case class DayHeadline(day: LocalDate) extends DateHeadline {
-  override val dateFormatString: String = "d MMMM yyyy"
-  override val dateTimeFormatString: String = "yyyy-MM-dd"
-  override val urlFragmentFormatString: Option[String] = Some("yyyy/MMM/dd")
-}
-
-case class MonthHeadline(day: LocalDate) extends DateHeadline {
-  override val dateFormatString: String = "MMMM yyyy"
-  override val dateTimeFormatString: String = "yyyy-MM"
-  override val urlFragmentFormatString: Option[String] = None
 }
