+++ common/app/layout/SliceWithCards.scala
@@ -103,5 +103,3 @@ object SliceWithCards {
     (SliceWithCards(layout.cssClassName, columns), unconsumed, endContext)
   }
 }
-
-
