+++ common/app/layout/Sublinks.scala
@@ -4,17 +4,6 @@ import cards._
 import implicits.FaciaContentFrontendHelpers.FaciaContentFrontendHelper
 import model.pressed.PressedContent
 
-object InclusiveRange {
-  def unit(n: Int): InclusiveRange = InclusiveRange(n, n)
-
-  def fromZero(n: Int): InclusiveRange = InclusiveRange(0, n)
-
-  def constrain(range: InclusiveRange, n: Int): Int =
-    (n max range.minimum) min range.maximum
-}
-
-case class InclusiveRange(minimum: Int, maximum: Int)
-
 /** The item class determines how many sublinks it has space to display. */
 object Sublinks {
   import InclusiveRange.{unit, fromZero}
