+++ common/app/layout/slices/ContainerDefinition.scala
@@ -2,11 +2,6 @@ package layout.slices
 
 import model.pressed.PressedContent
 
-sealed trait MobileShowMore
-
-case object DesktopBehaviour extends MobileShowMore
-case class RestrictTo(items: Int) extends MobileShowMore
-
 object ContainerDefinition {
   val DefaultCards = 6
