+++ common/app/layout/slices/ContainerJsonConfig.scala
@@ -2,11 +2,11 @@ package layout.slices
 
 import play.api.libs.json.Json
 
-object ContainerJsonConfig {
-  implicit val jsonFormat = Json.format[ContainerJsonConfig]
-}
-
 case class ContainerJsonConfig(
   name: String,
   groups: Option[Seq[String]]
 )
+
+object ContainerJsonConfig {
+  implicit val jsonFormat = Json.format[ContainerJsonConfig]
+}
