+++ common/app/layout/slices/DynamicContainers.scala
@@ -1,0 +1,21 @@
+package layout.slices
+
+import conf.Configuration
+import model.pressed.PressedContent
+
+object DynamicContainers {
+  val all: Map[String, DynamicContainer] = Map(
+    ("dynamic/fast", DynamicFast),
+    ("dynamic/slow", DynamicSlow),
+    ("dynamic/package", DynamicPackage),
+    ("dynamic/slow-mpu", DynamicSlowMPU(omitMPU = false, adFree = false))
+  )
+
+  def apply(collectionType: Option[String], items: Seq[PressedContent]): Option[ContainerDefinition] = {
+    for {
+      typ <- collectionType
+      dynamicContainer <- all.get(typ)
+      definition <- dynamicContainer.containerDefinitionFor(items.map(Story.fromFaciaContent))
+    } yield definition
+  }
+}
