+++ common/app/layout/slices/MobileShowMore.scala
@@ -1,0 +1,8 @@
+package layout.slices
+
+import model.pressed.PressedContent
+
+sealed trait MobileShowMore
+
+case object DesktopBehaviour extends MobileShowMore
+case class RestrictTo(items: Int) extends MobileShowMore
